Created split.py
The split.py script allows splitting a txt file of a book into pages (note: in lines 48 and 49, the correct marker for the pages within the book.txt file should be used instead of "Page" and "Page | 31").
It also encodes the single lines using cryptography package's Fernet before writing them to a different file.